Now using systemd instead of forever
Added Dockerfile and .dockerignore and have successfully built a docker
image and run the container with Slack integration. Moved away from
using forever, now running as a service using systemd. This is done in a
Docker container mostly to make it easy for someone to try it out. The
service could be installed directly instead of running it its own Docker
container.