Very basic working concept
Have the main app try to start Slack integration when the app itself
starts, so that if a token has been previously saved it will start up on
its own on a restart. Created a form on the settings page to store a
"motion snapshot url" with a nickname and url for each snapshot entry.
Added settings functions to save the motion snapshot info. Slacks.js
listens for "show snapshot office" for now, and grabs the URL from the
levelDB data store associated with the name "office", and uploads it to
the correct Slack channel.